Center CT labels on canvas, truncate the label if necessary
